feat: 상품 등록 API의 상품 가격 저장 기능 구현

+++ src/main/java/com/wanted/cqrs/service/request/ProductServiceRequest.java
@@ -1,5 +1,6 @@
 package com.wanted.cqrs.service.request;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -41,11 +42,11 @@ public class ProductServiceRequest {
     @Builder
     public static class ProductPrice {
 
-        private Long basePrice;
-        private Long salePrice;
-        private Long costPrice;
+        private BigDecimal basePrice;
+        private BigDecimal salePrice;
+        private BigDecimal costPrice;
         private String currency;
-        private Integer taxRate;
+        private BigDecimal taxRate;
     }
 
     @Builder
